final_backend_ready/final_backend/src/utils/db_utils.py
import psycopg2

def get_database_connection():
    return psycopg2.connect(
        host="localhost",
        port="5432",
        database="SecureDocAI",
        user="postgres",
        password="5103"
    )